# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Konouz Islamic Books Encyclopedia API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Book {
  /** Archive.org item identifier */
  identifier: string;
  title: string;
  creator?: string | null;
  year?: string | null;
  description?: string | null;
  /** URL to cover thumbnail */
  thumbnail: string;
  /** Archive.org online reader URL */
  readUrl: string;
  /** Direct PDF download URL */
  downloadUrl: string;
  downloads?: number | null;
}

export interface SearchResults {
  query: string;
  page: number;
  total: number;
  books: Book[];
}

export interface PopularSearch {
  query: string;
  count: number;
}

export interface SearchStats {
  totalSearches: number;
  uniqueQueries: number;
  last24h: number;
}

export type SearchBooksParams = {
  /**
   * @minLength 1
   */
  q: string;
  /**
   * @minimum 1
   */
  page?: number;
  /**
   * @minimum 1
   * @maximum 50
   */
  rows?: number;
};