# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Konouz Islamic Books Encyclopedia API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /books/search:
    get:
      operationId: searchBooks
      tags: [books]
      summary: Search Islamic books on Archive.org
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: rows
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 24
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResults"
...
components:
  schemas:
...
    Book:
      type: object
      properties:
        identifier:
          type: string
          description: Archive.org item identifier
        title:
          type: string
        creator:
          type: string
          nullable: true
        year:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        thumbnail:
          type: string
          description: URL to cover thumbnail
        readUrl:
          type: string
          description: Archive.org online reader URL
        downloadUrl:
          type: string
          description: Direct PDF download URL
        downloads:
          type: integer
          nullable: true
      required: [identifier, title, thumbnail, readUrl, downloadUrl]
    SearchResults:
      type: object
      properties:
        query:
          type: string
        page:
          type: integer
        total:
          type: integer
        books:
          type: array
          items:
            $ref: "#/components/schemas/Book"
      required: [query, page, total, books]
...