# lib/api-zod/src/generated/types/book.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Konouz Islamic Books Encyclopedia API
 * OpenAPI spec version: 0.1.0
 */

export interface Book {
  /** Archive.org item identifier */
  identifier: string;
  title: string;
  creator?: string | null;
  year?: string | null;
  description?: string | null;
  /** URL to cover thumbnail */
  thumbnail: string;
  /** Archive.org online reader URL */
  readUrl: string;
  /** Direct PDF download URL */
  downloadUrl: string;
  downloads?: number | null;
}