# lib/api-zod/src/generated/types/searchResults.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Konouz Islamic Books Encyclopedia API
 * OpenAPI spec version: 0.1.0
 */
import type { Book } from "./book";

export interface SearchResults {
  query: string;
  page: number;
  total: number;
  books: Book[];
}